split up generic sub-package into separate modules
